CurrentRange_R3: allow touchpad setting of sample rate
Use "sample rate" instead of oversample rate to get a more natural
match to user expectations: increase should increase the sample rate,
which means decrease the oversample rate.

Add a preferences byte persisted in EEPROM.  Use this to record
whether the user wants simultaneous touch of uA+mA to toggle offset
(bias) mode (default) or to repurpose the nA and mA keys to increase
or decrease the sample rate.

Changes to sample rate through the serial interface are persisted to
EEPROM.

Changes to the sample rate from the touchpads are not persisted.  A
tone is emitted to audibly reflect the selected sample rate.

Signed-off-by: Peter A. Bigot <pab@pabigot.com>