CurrentRanger_R3: add display of configured logging format
The logging format is preserved in EEPROM, but never displayed, so you
can see the selected value only by changing it.  Add it to the set of
preferences that are displayed in the config/calib block.

Also start marking the key documentation for things that are
"preferences" that are saved between restarts.

Signed-off-by: Peter A. Bigot <pab@pabigot.com>