CurrentRanger_R3: add preference to timestamp samples
For USB and BT serial data output add options to preface each sample
with the system time at which it was taken, in microseconds.  This
mitigates the loss of information when the sample rate is changed
dynamically.

Be aware that the counter will wrap at 2^32 us, so relative
calculations should be done using unsigned 32-bit operations.

Signed-off-by: Peter A. Bigot <pab@pabigot.com>